Improve filter by games played and support ties
When processing all divisions in the AHF, it's impractical to call out
each showcase team.  Instead, filter out teams by number of games
played.

Originally the code was written with the assumption that due to the
shootouts, games cannot end in a tie.  It turns out some games this
season have ended in a tie, and this patch includes support for ties.