Add support for fake tie games
Some KRACH implementations have a potential divide-by-zero risk when
there is an undefeated team. To work around this, KRACH used to create a
fake team, and use it to give every real team a tie. Ties count as half
of a 'loss point', hence ensuring no divide by zero was possible.

This tool does not have any divide-by-zero risk, but having this as an
option gives us another knob for trying to reproduce results from other
leagues.