AHF: update to treat shootouts as win-loss
Previously, the AHF told me they treated shootouts as a normal win-loss,
but this tool produced better results when treating all shootouts as a
tie.

Turns out that the AHF was mis-calculating their numbers, and really
were treating shootouts as a tie. Word is that they will be updating
to calculate things "correctly".