krach.py: swap from num games to sum of win points
See comments in the patch, but there's some disagreement in the websites
about one term in the equation, whether it's "number of games played"
between two opponents, versus "sum of win points".

For most situations, they are equivalent because every game is worth 1.0
win point in total.  Even in a tie, each team still gets 0.5 win points
to maintain the 1.0 total.

But the AHF uses 0.85 win points for a tie (at least for a tie in one of
their 'dummy' games).  And because both teams earn that amount the total
win points for those games is actually 1.70.

Shrug ... switch to "sum of win points" to support all cases.