Cache divisions in local file
Previously every time a command was run (download, update, teams) it
would make an API query to get the current list of divisions.

THis patch stores off the divisions in a local file.  Next time a
command is run, use the cache instead of download from the
gamesheetstats API.

To force a re-download of the division, delete the cached file.