Disabled scheduled updateds in github CI
API request for scores has started to fail; Disable github CI to avoid
noise of the nightly updates failing.

NOTE: I have not tried to debug the API failure; there was recently some
tension between black bear and my hockey rankings over API access to the
game data.  I didn't really follow along, but possible that this API
failure is in some way an effort by gamesheets to crack down on raw
data access.  Or it could be a natural evolution of the API, and my
current code is using a defunct/deprecated API.